+++ terraform/modules/fargate-task/iam.tf
@@ -21,7 +21,7 @@ resource "aws_iam_role_policy_attachment" "ecs_task_execution_role_policy" {
 }
 
 resource "aws_iam_role" "ecs_task_role" {
-  name_prefix = "${local.workspace_prefix}-ecs-role-"
+  name_prefix = "${local.workspace_prefix}-ecs-task-role-"
 
   assume_role_policy = jsonencode({
     Version = "2012-10-17",
@@ -35,6 +35,10 @@ resource "aws_iam_role" "ecs_task_role" {
       }
     ]
   })
+
+  tags = {
+    TaskName = var.task_name
+  }
 }
 
 
@@ -58,16 +62,22 @@ resource "aws_iam_policy" "s3_read_write_policy" {
           "s3:List*"
         ],
         Resource = [
-          "arn:aws:s3:::sfc-${local.workspace_prefix}-datasets/*",
-          "arn:aws:s3:::sfc-${local.workspace_prefix}-datasets"
+          "arn:aws:s3:::sfc-${terraform.workspace}-datasets/*",
+          "arn:aws:s3:::sfc-${terraform.workspace}-datasets",
+          "arn:aws:s3:::sfc-${terraform.workspace}-pipeline-resources/*",
+          "arn:aws:s3:::sfc-${terraform.workspace}-pipeline-resources"
         ]
       }
     ]
   })
 }
 
+data "aws_secretsmanager_secret" "cqc_api_primary_key" {
+  name = var.secret_name
+}
+
 resource "aws_iam_policy" "secretsmanager_read_policy" {
-  name_prefix = "${local.workspace_prefix}-secretsmanager-read-policy-"
+  name_prefix = "${local.workspace_prefix}-secretsmanager-"
   description = "IAM policy for Secrets Manager read access to a specific secret."
   policy = jsonencode({
     Version = "2012-10-17",
@@ -78,10 +88,14 @@ resource "aws_iam_policy" "secretsmanager_read_policy" {
         Action = [
           "secretsmanager:GetSecretValue"
         ],
-        Resource = "arn:aws:secretsmanager:${var.region}:${local.account_id}:secret:${var.secret_arn_label}",
+        Resource = data.aws_secretsmanager_secret.cqc_api_primary_key.arn
       }
     ]
   })
+
+  tags = {
+    TaskName = var.task_name
+  }
 }
 
 resource "aws_iam_role_policy_attachment" "ecs_task_role_policy_read_secr" {
@@ -89,8 +103,100 @@ resource "aws_iam_role_policy_attachment" "ecs_task_role_policy_read_secr" {
   policy_arn = aws_iam_policy.secretsmanager_read_policy.arn
 }
 
+resource "aws_iam_policy" "ssm_parameter_store_policy" {
+  name_prefix = "SSMParamStorePolicyModels-"
+  path        = "/"
+  description = "A broadly permissive policy for AWS Systems Manager Parameter Store."
+
+  policy = <<EOF
+{
+    "Version": "2012-10-17",
+    "Statement": [
+        {
+            "Sid": "ReadAccess",
+            "Effect": "Allow",
+            "Action": [
+                "ssm:GetParameter",
+                "ssm:GetParameters",
+                "ssm:GetParametersByPath",
+                "ssm:DescribeParameters"
+            ],
+            "Resource": "*"
+        },
+        {
+            "Sid": "WriteAccess",
+            "Effect": "Allow",
+            "Action": [
+                "ssm:PutParameter",
+                "ssm:DeleteParameter",
+                "ssm:LabelParameterVersion"
+            ],
+            "Resource": "*"
+        }
+    ]
+}
+EOF
+}
+
+resource "aws_iam_role_policy_attachment" "ecs_task_role_policy_params" {
+  role       = aws_iam_role.ecs_task_role.name
+  policy_arn = aws_iam_policy.ssm_parameter_store_policy.arn
+}
+
+resource "aws_iam_policy" "sns_publish_policy" {
+  name_prefix = "SNSPublishPolicyModels-"
+  path        = "/"
+  description = "A policy allowing publication to a specific SNS topic."
+
+  policy = <<EOF
+{
+    "Version": "2012-10-17",
+    "Statement": [
+        {
+            "Sid": "PublishToSNS",
+            "Effect": "Allow",
+            "Action": ["sns:Publish", "sns:GetTopicAttributes", "sns:SetTopicAttributes", "sns:TagResource"],
+            "Resource": "*"
+        }
+    ]
+}
+EOF
+}
+
+resource "aws_iam_role_policy_attachment" "ecs_task_role_policy_sns" {
+  role       = aws_iam_role.ecs_task_role.name
+  policy_arn = aws_iam_policy.sns_publish_policy.arn
+}
+
+data "aws_iam_policy_document" "step_func_success_failure" {
+  statement {
+    effect = "Allow"
+
+    actions = [
+      "states:SendTaskSuccess",
+      "states:SendTaskFailure"
+    ]
+
+    resources = [
+      "*"
+    ]
+  }
+}
+
+resource "aws_iam_policy" "sfn_task_policy" {
+  name_prefix = "StepFunctionActions-"
+  policy      = data.aws_iam_policy_document.step_func_success_failure.json
+  path        = "/"
+  description = "A policy allowing success or failure notifications to be sent to SFN."
+}
+
+resource "aws_iam_role_policy_attachment" "stepfn_policy_attach" {
+  policy_arn = aws_iam_policy.sfn_task_policy.arn
+  role       = aws_iam_role.ecs_task_role.name
+}
+
 resource "aws_iam_role" "sfn_execution_role" {
-  name_prefix = "${local.workspace_prefix}-sfn-exec-role-"
+  name_prefix = "${local.workspace_prefix}-sfn-"
 
   assume_role_policy = jsonencode({
     Version = "2012-10-17",
@@ -104,10 +210,14 @@ resource "aws_iam_role" "sfn_execution_role" {
       }
     ]
   })
+
+  tags = {
+    TaskName = var.task_name
+  }
 }
 
 resource "aws_iam_policy" "sfn_ecs_policy" {
-  name_prefix = "${local.workspace_prefix}-sfn-ecs-policy-"
+  name_prefix = "${local.workspace_prefix}-${var.task_name}-sfn-"
   policy = jsonencode({
     Version = "2012-10-17",
     Statement = [