+++ terraform/pipeline/main.tf
@@ -15,7 +15,7 @@ provider "aws" {
 
 terraform {
   backend "s3" {
-    bucket         = "sfc-terraform-state"
+    # Bucket defined in ../*.s3.tfbackend
     key            = "statefiles/workspace=default/terraform.tfstate"
     region         = "eu-west-2"
     dynamodb_table = "terraform-locks"
@@ -24,9 +24,8 @@ terraform {
 }
 
 locals {
-  workspace_prefix               = substr(lower(replace(terraform.workspace, "/[^a-zA-Z0-9]+/", "-")), 0, 30)
-  is_development_environment     = local.workspace_prefix != "main"
-  cqc_api_primary_key_secret_arn = "arn:aws:secretsmanager:${var.region}:${data.aws_caller_identity.current.account_id}:secret:cqc_api_primary_key-mK4hzZ"
+  workspace_prefix           = substr(lower(replace(terraform.workspace, "/[^a-zA-Z0-9]+/", "-")), 0, 30)
+  is_development_environment = local.workspace_prefix != "main"
 }
 
 data "aws_caller_identity" "current" {}