+++ tests/test_polars_utils/test_utils.py
@@ -10,7 +10,12 @@ from glob import glob
 from pathlib import Path
 from unittest.mock import patch
 
+import boto3
 import polars as pl
+from botocore.exceptions import ClientError
+from moto import mock_aws, sns
+from moto.core import DEFAULT_ACCOUNT_ID, set_initial_no_auth_action_count
+
 from polars_utils import utils
 from polars_utils.utils import write_to_parquet
 
@@ -253,3 +258,74 @@ class TestEmptyS3Folder(unittest.TestCase):
                 ]
             },
         )
+
+
+class SendSnsNotificationTests(TestUtils):
+    @mock_aws
+    def test_send_sns_notification_sends(self):
+        client = boto3.client("sns", region_name="eu-west-2")
+        topic = client.create_topic(Name="test-topic")
+
+        utils.send_sns_notification(
+            topic_arn=topic["TopicArn"],
+            subject="Test Subject",
+            message="Test Message",
+        )
+
+        sns_backend = sns.sns_backends[DEFAULT_ACCOUNT_ID]["eu-west-2"]
+        sent_notifications = sns_backend.topics[topic["TopicArn"]].sent_notifications
+
+        assert len(sent_notifications) == 1
+        notification = sent_notifications[0]
+        assert notification[2] == "Test Subject"
+        assert notification[1] == "Test Message"
+
+    @mock_aws
+    def test_send_sns_notification_raises_clienterror_and_logs_with_wrong_arn(self):
+        client = boto3.client("sns", region_name="eu-west-2")
+        topic = client.create_topic(Name="test-topic")
+        with self.assertLogs(level="ERROR") as cm:
+            with self.assertRaises(ClientError):
+                utils.send_sns_notification(
+                    topic_arn="silly_arn",
+                    subject="Test Subject",
+                    message="Test Message",
+                )
+        self.assertIn(
+            "There was an error writing to SNS - check your IAM permissions and that you have the right topic ARN",
+            cm.output[1],
+        )
+
+    @mock_aws
+    @set_initial_no_auth_action_count(1)
+    def test_send_sns_notification_raises_clienterror_and_logs_with_no_permissions(
+        self,
+    ):
+        client = boto3.client("sns", region_name="eu-west-2")
+        topic = client.create_topic(Name="test-topic")
+        with self.assertLogs(level="ERROR") as cm:
+            with self.assertRaises(ClientError):
+                utils.send_sns_notification(
+                    topic_arn=topic["TopicArn"],
+                    subject="Test Subject",
+                    message="Test Message",
+                )
+        self.assertIn(
+            "There was an error writing to SNS - check your IAM permissions and that you have the right topic ARN",
+            cm.output[1],
+        )
+
+    def test_parse_args_converts_simple_data_types(self):
+        self.assertTrue(utils.parse_arg_by_type("tRue"))
+        self.assertFalse(utils.parse_arg_by_type("fAlSe "))
+        self.assertEqual(19, utils.parse_arg_by_type("19 "))
+        self.assertEqual(19, utils.parse_arg_by_type(" 19"))
+        self.assertEqual(23.7, utils.parse_arg_by_type("23.7 "))
+        self.assertEqual(
+            "Dave is 173.4 cm tall", utils.parse_arg_by_type("Dave is 173.4 cm tall")
+        )
+
+    def test_parse_args_converts_complex_types(self):
+        self.assertEqual("[1, 2, 3]", utils.parse_arg_by_type("[1, 2, 3]"))
+        self.assertEqual("{1:2, 3:4}", utils.parse_arg_by_type("{1:2, 3:4}"))
+        self.assertEqual("2025-06-19", utils.parse_arg_by_type("2025-06-19"))